Productos model modificado

+++ backend/models/productos.model.js
@@ -8,12 +8,12 @@ module.exports = (sequelize, Sequelize) => {
             type: Sequelize.STRING,
             allowNull: false
         },
-        precio_venta: {
-            type: Sequelize.Float,
-            allowNull: false
-        },
+      precio_venta: {
+      type: Sequelize.DECIMAL(10, 2), // hasta 10 dígitos con 2 decimales
+      allowNull: false
+    },
         precio_compra: {
-            type: Sequelize.Float,
+            type: Sequelize.DECIMAL(10,2),
             allowNull: false
         }
     }, {
